Import of JESSE/Jifty-0.51228 from CPAN.
gitpan-cpan-distribution: Jifty
gitpan-cpan-version:      0.51228
gitpan-cpan-path:         JESSE/Jifty-0.51228.tar.gz
gitpan-cpan-author:       JESSE
gitpan-cpan-maturity:     released

+++ lib/Jifty/Test.pm
@@ -43,11 +43,13 @@ sub setup {
 
     Jifty->new( no_handle => 1 );
 
+    Log::Log4perl->get_logger("SchemaTool")->less_logging(3);
     my $schema = Jifty::Script::Schema->new;
     $schema->{drop_database} =
       $schema->{create_database} =
         $schema->{create_all_tables} = 1;
     $schema->run;
+    Log::Log4perl->get_logger("SchemaTool")->more_logging(3);
 
     Jifty->new();
     $class->setup_mailbox;
@@ -95,6 +97,7 @@ sub make_server {
     require Test::HTTP::Server::Simple;
     unshift @Jifty::Server::ISA, 'Test::HTTP::Server::Simple';
 
+    Log::Log4perl->get_logger("Jifty::Server")->less_logging(3);
     my $server = Jifty::Server->new;
 
     return $server;
@@ -156,7 +159,23 @@ sub messages {
 } 
 
 END {
-    unlink mailbox();
+    my $Test = Jifty::Test->builder;
+    # Such a hack -- try to detect if this is a forked copy and don't
+    # do cleanup in that case.
+    return if $Test->{Original_Pid} != $$;
+
+    # If all tests passed..
+    unless (grep {not $_} $Test->summary) {
+        # Clean up mailbox
+        unlink mailbox();
+
+        # Remove testing db
+        Log::Log4perl->get_logger("SchemaTool")->less_logging(3);
+        my $schema = Jifty::Script::Schema->new;
+        $schema->{drop_database} = 1;
+        $schema->run;
+        Log::Log4perl->get_logger("SchemaTool")->more_logging(3);
+    }
 }
 
 1;
